+++ src/nypl_py_utils/functions/config_helper.py
@@ -10,14 +10,13 @@ logger = create_log('config_helper')
 
 def load_env_file(run_type, file_string):
     """
-    This method loads a YAML config file containing environment variables,
-    decrypts whichever are encrypted, and puts them all into os.environ as
-    strings. For a YAML variable containing a list of values, the list is
-    exported into os.environ as a json string and should be loaded as such.
+    This method reads a YAML config file containing environment variables and
+    loads them all into os.environ as strings. See _parse_yaml_dict for more.
 
-    It requires the YAML file to be split into a 'PLAINTEXT_VARIABLES' section
-    and an 'ENCRYPTED_VARIABLES' section. See config/sample.yaml for an example
-    config file.
+    If the config file is divided into 'PLAINTEXT_VARIABLES' and
+    'ENCRYPTED_VARIABLES' sections (see config/sample.yaml for an exmaple), the
+    'ENCRYPTED_VARIABLES' variables will be decrypted first. Otherwise, all
+    variables will be loaded as is.
 
     Parameters
         ----------
@@ -36,31 +35,50 @@ def load_env_file(run_type, file_string):
             try:
                 env_dict = yaml.safe_load(env_stream)
             except yaml.YAMLError:
-                logger.error('Invalid YAML file: {}'.format(open_file))
                 raise ConfigHelperError(
                     'Invalid YAML file: {}'.format(open_file)) from None
     except FileNotFoundError:
-        logger.error('Could not find config file {}'.format(open_file))
         raise ConfigHelperError(
             'Could not find config file {}'.format(open_file)) from None
 
     if env_dict:
-        for key, value in env_dict.get('PLAINTEXT_VARIABLES', {}).items():
-            if type(value) is list:
-                os.environ[key] = json.dumps(value)
-            else:
-                os.environ[key] = str(value)
+        if ('PLAINTEXT_VARIABLES' in env_dict
+                or 'ENCRYPTED_VARIABLES' in env_dict):
+            _parse_yaml_dict(env_dict.get('PLAINTEXT_VARIABLES', {}))
 
-        kms_client = KmsClient()
-        for key, value in env_dict.get('ENCRYPTED_VARIABLES', {}).items():
-            if type(value) is list:
-                decrypted_list = [kms_client.decrypt(v) for v in value]
-                os.environ[key] = json.dumps(decrypted_list)
-            else:
-                os.environ[key] = kms_client.decrypt(value)
-        kms_client.close()
+            kms_client = KmsClient()
+            _parse_yaml_dict(env_dict.get(
+                'ENCRYPTED_VARIABLES', {}), kms_client)
+            kms_client.close()
+        else:
+            _parse_yaml_dict(env_dict)
+
+
+def _parse_yaml_dict(yaml_dict, kms_client=None):
+    """
+    Loads YAML dict into os.environ. All values are stored as strings to match
+    how AWS Lambda environment variables are stored. For list variables, the
+    list is exported into os.environ as a json string.
+
+    If kms_client is not empty, decrypts the variables first.
+
+    Does not allow for sub-dictionaries.
+    """
+    for key, value in yaml_dict.items():
+        if type(value) is dict:
+            raise ConfigHelperError(
+                'Found sub-dictionary in YAML config') from None
+        elif type(value) is list:
+            val = [kms_client.decrypt(v)
+                   for v in value] if kms_client else value
+            os.environ[key] = json.dumps(val)
+        else:
+            val = kms_client.decrypt(value) if kms_client else value
+            os.environ[key] = str(val)
 
 
 class ConfigHelperError(Exception):
     def __init__(self, message=None):
         self.message = message
+        if message is not None:
+            logger.error(message)