+++ tests/test_log_helper.py
@@ -11,13 +11,14 @@ from nypl_py_utils.functions.log_helper import create_log
 @freeze_time('2023-01-01 19:00:00')
 class TestLogHelper:
     def test_json_logging(self, capsys):
-        logger = create_log('test_log', json=True)
+        logger = create_log('test_structlog', json=True)
         logger.info('test', some="json")
         output = json.loads(capsys.readouterr().out)
         assert output.get("message") == 'test'
         assert output.get("some") == 'json'
         assert output.get('level') == 'info'
         assert output.get('timestamp') == '2023-01-01T19:00:00Z'
+        assert output.get('logger') == 'test_structlog'
 
     def test_default_logging(self, caplog):
         logger = create_log('test_log')